requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
cramjam==2.9.1
fastparquet==2024.11.0
fsspec==2025.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mysql-connector-python==9.2.0
numpy==1.26.4
ollama==0.4.7
packaging==24.2
pandas==1.5.3
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
